Introduce framework module namespaces and compatibility wiring
