+++ shared/src/commonMain/kotlin/com/devil/phoenixproject/framework/ui/README.kt
@@ -1,0 +1,8 @@
+@file:Suppress("unused")
+
+package com.devil.phoenixproject.framework.ui
+
+/**
+ * Placeholder namespace for reusable UI screens/components shared across products.
+ */
+internal object FrameworkUiNamespace
